refactor(ui): fast-refresh compliance batch 2 — router.tsx + __root.tsx
Part of #7850. 7 of 246 react-refresh/only-export-components warnings.

TanStack Router's convention (export const Route = createRootRouteWithContext()({component: X, ...}))
means a route function referenced from an exported Route config counts as
"exported" for fast-refresh purposes even without its own `export`
keyword, so router.tsx and __root.tsx defining their fallback/shell/root
components inline alongside the router/route config they belong to
tripped the rule on every one of those functions.

- router.tsx: moved DefaultRouteError/DefaultRoutePending to a new
  router-fallbacks.tsx sibling; getRouter() is now the file's only export.
- __root.tsx: moved NotFoundComponent/ErrorComponent/RootShell/
  RootComponent/RouteTransitionBar to a new -root-views.tsx sibling
  (`-`-prefixed so TanStack's file-based router ignores it, matching the
  existing routeFileIgnorePrefix convention); Route is now the file's
  only export. RootComponent used to read `const { queryClient } =
  Route.useRouteContext()` via same-module closure over Route -- moving
  it out of __root.tsx would make that circular (the route config needs
  RootComponent to construct Route, RootComponent needs Route to read
  context), so it switched to the equivalent standalone hook,
  `useRouteContext({ from: "__root__" })`, same runtime behavior without
  the circular import.

Zero behavior change; no external importers of any of the 7 moved
functions existed outside their own file, so this is fully mechanical.

246 -> 239 warnings; 0 eslint errors. Live-verified: homepage and a
missing route both render correctly, no console errors.
